simplify: per-strategy metric, trim narrative comments, explicit drops
clean.py candidates CTE now computes only the metric the chosen
overlap_strategy ranks by, instead of always computing both ST_Area
and ST_Length(ST_Intersection(boundaries)). The latter is the most
expensive op in the CTE — skipping it under the default `largest_area`
strategy is a real win on dirty data.

Drop loop in clean.py replaced with explicit per-table drops to match
the rest of app/.

merge.py polygonize comment: rewrote the stale "ST_Difference noise"
narration — clean.py uses polygonize, not ST_Difference. The actual
mechanism: polygonize adds crossing vertices on modified polygons that
don't appear in the points used to compute _05_tmp2's extension
boundary, so _02a closes the cell from the polygon's own boundary.

merge.py fallback comment: trim the change-history narration about the
prior CROSS JOIN OOM, keep the load-bearing complexity claim.

voronoi.py: drop redundant ::DOUBLE casts on ST_X/ST_Y (already DOUBLE).

Trim WHAT-comments in clean.py CTE walkthrough and tracker block.

Verified: all three datasets still pass the row-count check end-to-end.

+++ app/voronoi.py
@@ -21,8 +21,7 @@ def main(conn: DuckDBPyConnection, name: str) -> None:
         WITH unique_pts AS (
             SELECT geom FROM "{name}_03b"
             QUALIFY ROW_NUMBER() OVER (
-                PARTITION BY round(ST_X(geom)::DOUBLE, 12),
-                             round(ST_Y(geom)::DOUBLE, 12)
+                PARTITION BY round(ST_X(geom), 12), round(ST_Y(geom), 12)
             ) = 1
         )
         SELECT UNNEST(ST_Dump(
